+++ factcheck/cmd/api/di/provider.go
@@ -5,9 +5,9 @@ import (
 
 	"github.com/google/wire"
 
-	"github.com/kaogeek/line-fact-check/factcheck/cmd/api/config"
 	"github.com/kaogeek/line-fact-check/factcheck/cmd/api/internal/handler"
 	"github.com/kaogeek/line-fact-check/factcheck/cmd/api/internal/server"
+	"github.com/kaogeek/line-fact-check/factcheck/internal/config"
 	"github.com/kaogeek/line-fact-check/factcheck/internal/di"
 )
 
@@ -17,6 +17,7 @@ var ProviderSet = wire.NewSet(
 	di.ProviderSetDatabase,
 	di.ProviderSetRepo,
 	di.ProviderSetCore,
+	di.New,
 	handler.New,
 	server.New,
 	New,
