+++ factcheck/internal/di/wire_gen.go
@@ -7,7 +7,7 @@
 package di
 
 import (
-	"github.com/kaogeek/line-fact-check/factcheck/cmd/api/config"
+	"github.com/kaogeek/line-fact-check/factcheck/internal/config"
 	"github.com/kaogeek/line-fact-check/factcheck/internal/core"
 	"github.com/kaogeek/line-fact-check/factcheck/internal/data/postgres"
 	"github.com/kaogeek/line-fact-check/factcheck/internal/repo"
